The moral of the story: don't mix Node and Control
Node2d does not beling inside Control and has bugs. In this case, the
bug was that if the Control started out as hidden, Node2d would never
get events from it. I still think this is a bug and should be fixed,
but the hack of hiding invisible buttons on every system is fine.