Towards solving the peksy player + dead player in system issues
The player class now owns a lot less of the process of being a player.
Client decides when it's in control/etc. If we pull out a bit more we
might be able to unite the class with the Alien class, which would be
sweet. The goal is to give client control and let the client decide
which inventory counts.